--- service/src/main/java/bio/terra/workspace/service/resource/controlled/flight/clone/workspace/AwaitCloneReferenceResourceFlightStep.java
@@ -1,25 +1,0 @@
-package bio.terra.workspace.service.resource.controlled.flight.clone.workspace;
-
-import bio.terra.stairway.FlightContext;
-import bio.terra.stairway.Step;
-import bio.terra.stairway.StepResult;
-import bio.terra.stairway.exception.RetryException;
-
-public class AwaitCloneReferenceResourceFlightStep implements Step {
-
-  private final String flightId;
-
-  public AwaitCloneReferenceResourceFlightStep(String flightId) {
-    this.flightId = flightId;
-  }
-
-  @Override
-  public StepResult doStep(FlightContext context) throws InterruptedException, RetryException {
-    return null;
-  }
-
-  @Override
-  public StepResult undoStep(FlightContext context) throws InterruptedException {
-    return null;
-  }
-}
